refactored static function + docstring
