+++ geofetch/utils.py
@@ -6,6 +6,9 @@ import subprocess
 import sys
 import re
 import requests
+from io import StringIO
+import csv
+from typing import NoReturn, Dict, List, Union
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -19,7 +22,7 @@ URL_BY_ACC = {
 }
 
 
-def is_known_type(accn=None, typename=None):
+def is_known_type(accn: str = None, typename: str = None):
     """
     Determine if the given accession is of a known type.
 
@@ -49,7 +52,7 @@ def parse_accessions(input_arg, metadata_folder, just_metadata=False):
     interested in from that GSE#. An empty sample list means we should get all
     samples from that GSE#. This loop will create this dict.
 
-    :param input_arg:
+    :param input_arg: Input argument (GSE, or file)
     :param str metadata_folder: path to folder for accession metadata
     :param bool just_metadata: whether to only process metadata, not the
         actual data associated with the accession
@@ -119,13 +122,12 @@ def parse_accessions(input_arg, metadata_folder, just_metadata=False):
     return acc_GSE_list
 
 
-def parse_SOFT_line(l):
+def parse_SOFT_line(l: str) -> dict:
     """
     Parse SOFT formatted line, returning a dictionary with the key-value pair.
 
     :param str l: A SOFT-formatted line to parse ( !key = value )
     :return dict[str, str]: A python Dict object representing the key-value.
-    :raise InvalidSoftLineException: if given line can't be parsed as SOFT line
     """
     elems = l[1:].split("=")
     return {elems[0].rstrip(): elems[1].lstrip()}
@@ -134,7 +136,7 @@ def parse_SOFT_line(l):
 class AccessionException(Exception):
     """Exceptional condition(s) dealing with accession number(s)."""
 
-    def __init__(self, reason=""):
+    def __init__(self, reason: str = ""):
         """
         Optionally provide explanation for exceptional condition.
 
@@ -230,8 +232,12 @@ class Accession(object):
         return result_list
 
     @staticmethod
-    def _validate(accn):
-        """Determine if given value looks like an accession."""
+    def _validate(accn: str):
+        """
+        Determine if given value looks like an accession.
+        :param str accn: ordinary accession identifier.
+        :return: typename, number
+        """
         typename, number = split_accn(accn)
         if len(typename) != 3:
             raise AccessionException(
@@ -248,7 +254,7 @@ class Accession(object):
         return typename, number
 
     @staticmethod
-    def accn_type_exception(accn, typename, include_known=True):
+    def accn_type_exception(accn: str, typename: str, include_known: bool = True):
         """
         Create an exception instance based on an accession and a
         parsed unknown typename.
@@ -266,7 +272,7 @@ class Accession(object):
         return AccessionException(message)
 
 
-def split_accn(accn):
+def split_accn(accn: str):
     """
     Split accession into prefix and number, leaving suffix as text
     and converting the type prefix to uppercase.
@@ -341,4 +347,324 @@ def run_subprocess(*args, **kwargs):
         sys.exit(1)
 
 
-# file_gse_content
+def _get_list_of_keys(list_of_dict: list):
+    """
+    Getting list of all keys that are in the dictionaries in the list
+
+    :param list list_of_dict: list of dicts with metadata
+    :return list: list of dictionary keys
+    """
+
+    list_of_keys = []
+    for element in list_of_dict:
+        list_of_keys.extend(list(element.keys()))
+    return list(set(list_of_keys))
+
+
+def _get_value(all_line: str):
+    """
+    :param all_line: string with key value. (e.g. '!Series_geo_accession = GSE188720')
+    :return: value (e.g. GSE188720)
+    """
+    line_value = all_line.split("= ")[-1]
+    return line_value.split(": ")[-1].rstrip("\n")
+
+
+def _read_tar_filelist(raw_text: str) -> dict:
+    """
+    Creating list for supplementary files that are listed in "filelist.txt"
+    :param str raw_text: path to the file with information about files that are zipped ("filelist.txt")
+    :return dict: dict of supplementary file names and additional information
+    """
+    f = StringIO(raw_text)
+    files_info = {}
+    csv_reader = csv.reader(f, delimiter="\t")
+    line_count = 0
+    for row in csv_reader:
+        if line_count == 0:
+            name_index = row.index("Name")
+            size_index = row.index("Size")
+            type_index = row.index("Type")
+
+            line_count += 1
+        else:
+            files_info[row[name_index]] = {
+                "file_size": row[size_index],
+                "type": row[type_index],
+            }
+
+    return files_info
+
+
+def _check_file_existance(meta_processed_sample: list) -> list:
+    """
+    Checking if last element of the list has files. If list of files is empty deleting it
+    :param: meta_processed_sample: list with metadata dictionary
+    :return: list with metadata dictionary after processing
+    """
+    nb = len(meta_processed_sample) - 1
+    if nb > -1:
+        if len(meta_processed_sample[nb]["files"]) == 0:
+            del meta_processed_sample[nb]
+            nb -= 1
+    return meta_processed_sample
+
+
+def _separate_list_of_files(meta_list: Union[list, dict], col_name: str = "files"):
+    """
+    This method is separating list of files (dict value) or just simple dict
+    into two different dicts
+    :param col_name: column name that should be added with filenames
+    :param meta_list: list, or dict with metadata
+    """
+    separated_list = []
+    if isinstance(meta_list, list):
+        for meta_elem in meta_list:
+            for file_elem in meta_elem[col_name]:
+                new_dict = meta_elem.copy()
+                new_dict.pop(col_name, None)
+                new_dict["file"] = file_elem
+                separated_list.append(new_dict)
+    elif isinstance(meta_list, dict):
+        for file_elem in meta_list[col_name]:
+            new_dict = meta_list.copy()
+            new_dict.pop(col_name, None)
+            new_dict["file"] = file_elem
+            separated_list.append(new_dict)
+    else:
+        return TypeError("Incorrect type")
+
+    return separated_list
+
+
+def _update_columns(
+    metadata: dict, experiment_name: str, sample_name: str, read_type: str
+) -> dict:
+    """
+    Update the metadata associated with a particular experiment.
+
+    For the experiment indicated, this function updates the value (mapping),
+    including new data and populating columns used by looper based on
+    existing values in the mapping.
+
+    :param Mapping metadata: the key-value mapping to update
+    :param str experiment_name: name of the experiment from which these
+        data came and are associated; the key in the metadata mapping
+        for which the value is to be updated
+    :param str sample_name: name of the sample with which these data are
+        associated
+    :param str read_type: usually "single" or "paired," an indication of the
+        type of sequencing reads for this experiment
+    :return: updated metadata
+    """
+
+    exp = metadata[experiment_name]
+
+    # Protocol-agnostic
+    exp["sample_name"] = sample_name
+    exp["protocol"] = exp["Sample_library_selection"]
+    exp["read_type"] = read_type
+    exp["organism"] = exp["Sample_organism_ch1"]
+    exp["data_source"] = "SRA"
+    exp["SRX"] = experiment_name
+
+    # Protocol specified is lowercased prior to checking here to alleviate
+    # dependence on case for the value in the annotations file.
+    bisulfite_protocols = {"reduced representation": "RRBS", "random": "WGBS"}
+
+    # Conditional on bisulfite sequencing
+    # print(":" + exp["Sample_library_strategy"] + ":")
+    # Try to be smart about some library methods, refining protocol if possible.
+    if exp["Sample_library_strategy"] == "Bisulfite-Seq":
+        # print("Parsing protocol")
+        proto = exp["Sample_library_selection"].lower()
+        if proto in bisulfite_protocols:
+            exp["protocol"] = bisulfite_protocols[proto]
+
+    return exp
+
+
+def _sanitize_config_string(text: str) -> str:
+    """
+    Function that sanitizes text in config file.
+    :param text: Any string that have to be sanitized
+    :return: sanitized strings
+    """
+    new_str = text
+    new_str = new_str.replace('"', f'\\"')
+    new_str = new_str.replace("'", f"''")
+    return new_str
+
+
+def _sanitize_name(name_str: str) -> str:
+    """
+    Function that sanitizes strings. (Replace all odd characters)
+    :param str name_str: Any string value that has to be sanitized.
+    :return: sanitized strings
+    """
+    new_str = name_str
+    punctuation1 = r"""!"#$%&'()*,./:;<=>?@[\]^_`{|}~"""
+    for odd_char in list(punctuation1):
+        new_str = new_str.replace(odd_char, "_")
+    new_str = new_str.replace(" ", "_").replace("__", "_")
+    return new_str
+
+
+def _create_dot_yaml(file_path: str, yaml_path: str) -> NoReturn:
+    """
+    Function that creates .pep.yaml file that points to actual yaml file
+    :param str file_path: Path to the .pep.yaml file that we want to create
+    :param str yaml_path: path or name of the actual yaml file
+    """
+    with open(file_path, "w+") as file:
+        file.writelines(f"config_file: {yaml_path}")
+
+
+def _which(program: str):
+    """
+    return str:  the path to a program to make sure it exists
+    """
+    import os
+
+    def is_exe(fp):
+        return os.path.isfile(fp) and os.access(fp, os.X_OK)
+
+    fpath, fname = os.path.split(program)
+    if fpath:
+        if is_exe(program):
+            return program
+    else:
+        for path in os.environ["PATH"].split(os.pathsep):
+            path = path.strip('"')
+            exe_file = os.path.join(path, program)
+            if is_exe(exe_file):
+                return exe_file
+
+
+def _dict_to_list_converter(
+    proj_dict: Dict = None, proj_list: List = None
+) -> Union[Dict, List]:
+    """
+    Converter project dict to list and vice versa
+    dict -> list
+    list -> dict
+    :param proj_dict: project dictionary
+    :param proj_list: project list
+    :return: converted values
+    """
+    if proj_dict is not None:
+        new_meta_list = []
+        for key in proj_dict:
+            new_dict = proj_dict[key]
+            new_dict["big_key"] = key
+            new_meta_list.append(new_dict)
+
+        meta_list = new_meta_list
+
+    elif proj_list is not None:
+        new_sample_dict = {}
+        for sample in proj_list:
+            new_sample_dict[sample["big_key"]] = sample
+        meta_list = new_sample_dict
+
+    else:
+        raise ValueError
+
+    return meta_list
+
+
+def _standardize_colnames(meta_list: Union[list, dict]) -> Union[list, dict]:
+    """
+    Standardize column names by lower-casing and underscore
+    :param list meta_list: list of dictionaries of samples
+    :return : list of dictionaries of samples with standard colnames
+    """
+    # check if meta_list is dict and converting it to list
+    input_is_dict = False
+    if isinstance(meta_list, dict):
+        input_is_dict = True
+        meta_list = _dict_to_list_converter(proj_dict=meta_list)
+
+    new_metalist = []
+    list_keys = _get_list_of_keys(meta_list)
+    for item_nb, values in enumerate(meta_list):
+        new_metalist.append({})
+        for key in list_keys:
+            try:
+                new_key_name = key.lower().strip()
+                new_key_name = _sanitize_name(new_key_name)
+
+                new_metalist[item_nb][new_key_name] = values[key]
+
+            except KeyError:
+                pass
+
+    if input_is_dict:
+        new_metalist = _dict_to_list_converter(proj_list=new_metalist)
+
+    return new_metalist
+
+
+def _separate_file_url(meta_list):
+    """
+    This method is adding dict key without file_name without path
+    """
+    separated_list = []
+    for meta_elem in meta_list:
+        new_dict = meta_elem.copy()
+        new_dict["file_url"] = meta_elem["file"]
+        new_dict["file"] = os.path.basename(meta_elem["file"])
+        # new_dict["sample_name"] = os.path.basename(meta_elem["file"])
+        try:
+            new_dict["sample_name"] = str(meta_elem["Sample_title"])
+            if new_dict["sample_name"] == "" or new_dict["sample_name"] is None:
+                raise KeyError("sample_name Does not exist. Creating .. ")
+        except KeyError:
+            new_dict["sample_name"] = os.path.basename(meta_elem["file"])
+
+        # sanitize sample names
+        new_dict["sample_name"] = _sanitize_name(new_dict["sample_name"])
+
+        separated_list.append(new_dict)
+    return separated_list
+
+
+def _filter_gsm(meta_processed_samples: list, gsm_list: dict) -> list:
+    """
+    Getting metadata list of all samples of one experiment and filtering it
+    by the list of GSM that was specified in the input files.
+    And then changing names of the sample names.
+
+    :param meta_processed_samples: list of metadata dicts of samples
+    :param gsm_list: list of dicts where GSM (samples) are keys and
+        sample names are values. Where values can be empty string
+    """
+
+    if gsm_list.keys():
+        new_gsm_list = []
+        for gsm_sample in meta_processed_samples:
+            if gsm_sample["Sample_geo_accession"] in gsm_list.keys():
+                gsm_sample_new = gsm_sample
+                if gsm_list[gsm_sample["Sample_geo_accession"]] != "":
+                    gsm_sample_new["sample_name"] = gsm_list[
+                        gsm_sample["Sample_geo_accession"]
+                    ]
+                new_gsm_list.append(gsm_sample_new)
+        return new_gsm_list
+    return meta_processed_samples
+
+
+def _unify_list_keys(processed_meta_list: list) -> list:
+    """
+    Unifying list of dicts with metadata, so every dict will have
+        same keys
+
+    :param list processed_meta_list: list of dicts with metadata
+    :return list: list of unified dicts with metadata
+    """
+    list_of_keys = _get_list_of_keys(processed_meta_list)
+    for k in list_of_keys:
+        for list_elem in range(len(processed_meta_list)):
+            if k not in processed_meta_list[list_elem]:
+                processed_meta_list[list_elem][k] = ""
+    return processed_meta_list
